update grafana and prometheus
