--- controllers/argocd/ingress.go
@@ -1,185 +1,0 @@
-// Copyright 2019 ArgoCD Operator Developers
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// 	http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package argocd
-
-import (
-	"context"
-	"fmt"
-
-	networkingv1 "k8s.io/api/networking/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
-
-	argoproj "github.com/argoproj-labs/argocd-operator/api/v1beta1"
-	"github.com/argoproj-labs/argocd-operator/common"
-	"github.com/argoproj-labs/argocd-operator/pkg/argoutil"
-)
-
-// getArgoServerPath will return the Ingress Path for the Argo CD component.
-func getPathOrDefault(path string) string {
-	result := common.ArgoCDDefaultIngressPath
-	if len(path) > 0 {
-		result = path
-	}
-	return result
-}
-
-// newIngress returns a new Ingress instance for the given ArgoCD.
-func newIngress(cr *argoproj.ArgoCD) *networkingv1.Ingress {
-	return &networkingv1.Ingress{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      cr.Name,
-			Namespace: cr.Namespace,
-			Labels:    argoutil.LabelsForCluster(cr),
-		},
-	}
-}
-
-// newIngressWithName returns a new Ingress with the given name and ArgoCD.
-func newIngressWithName(name string, cr *argoproj.ArgoCD) *networkingv1.Ingress {
-	ingress := newIngress(cr)
-	ingress.ObjectMeta.Name = name
-
-	lbls := ingress.ObjectMeta.Labels
-	lbls[common.ArgoCDKeyName] = name
-	ingress.ObjectMeta.Labels = lbls
-
-	return ingress
-}
-
-// newIngressWithSuffix returns a new Ingress with the given name suffix for the ArgoCD.
-func newIngressWithSuffix(suffix string, cr *argoproj.ArgoCD) *networkingv1.Ingress {
-	return newIngressWithName(fmt.Sprintf("%s-%s", cr.Name, suffix), cr)
-}
-
-// reconcileIngresses will ensure that all ArgoCD Ingress resources are present.
-func (r *ReconcileArgoCD) reconcileIngresses(cr *argoproj.ArgoCD) error {
-	if err := r.reconcileArgoServerIngress(cr); err != nil {
-		return err
-	}
-
-	if err := r.reconcileArgoServerGRPCIngress(cr); err != nil {
-		return err
-	}
-
-	if err := r.reconcileGrafanaIngress(cr); err != nil {
-		return err
-	}
-
-	if err := r.reconcilePrometheusIngress(cr); err != nil {
-		return err
-	}
-
-	if err := r.reconcileApplicationSetControllerIngress(cr); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// reconcileGrafanaIngress will ensure that the ArgoCD Server GRPC Ingress is present.
-func (r *ReconcileArgoCD) reconcileGrafanaIngress(cr *argoproj.ArgoCD) error {
-	ingress := newIngressWithSuffix("grafana", cr)
-	if argoutil.IsObjectFound(r.Client, cr.Namespace, ingress.Name, ingress) {
-		if !cr.Spec.Grafana.Enabled || !cr.Spec.Grafana.Ingress.Enabled {
-			// Ingress exists but enabled flag has been set to false, delete the Ingress
-			return r.Client.Delete(context.TODO(), ingress)
-		}
-		log.Info(grafanaDeprecatedWarning)
-		return nil // Ingress found and enabled, do nothing
-	}
-
-	if !cr.Spec.Grafana.Enabled || !cr.Spec.Grafana.Ingress.Enabled {
-		return nil // Grafana itself or Ingress not enabled, move along...
-	}
-
-	log.Info(grafanaDeprecatedWarning)
-
-	return nil
-}
-
-// reconcilePrometheusIngress will ensure that the Prometheus Ingress is present.
-func (r *ReconcileArgoCD) reconcilePrometheusIngress(cr *argoproj.ArgoCD) error {
-	ingress := newIngressWithSuffix("prometheus", cr)
-	if argoutil.IsObjectFound(r.Client, cr.Namespace, ingress.Name, ingress) {
-		if !cr.Spec.Prometheus.Enabled || !cr.Spec.Prometheus.Ingress.Enabled {
-			// Ingress exists but enabled flag has been set to false, delete the Ingress
-			return r.Client.Delete(context.TODO(), ingress)
-		}
-		return nil // Ingress found and enabled, do nothing
-	}
-
-	if !cr.Spec.Prometheus.Enabled || !cr.Spec.Prometheus.Ingress.Enabled {
-		return nil // Prometheus itself or Ingress not enabled, move along...
-	}
-
-	// Add default annotations
-	atns := make(map[string]string)
-	atns[common.ArgoCDKeyIngressSSLRedirect] = "true"
-	atns[common.ArgoCDKeyIngressBackendProtocol] = "HTTP"
-
-	// Override default annotations if specified
-	if len(cr.Spec.Prometheus.Ingress.Annotations) > 0 {
-		atns = cr.Spec.Prometheus.Ingress.Annotations
-	}
-
-	ingress.ObjectMeta.Annotations = atns
-
-	ingress.Spec.IngressClassName = cr.Spec.Prometheus.Ingress.IngressClassName
-
-	pathType := networkingv1.PathTypeImplementationSpecific
-	// Add rules
-	ingress.Spec.Rules = []networkingv1.IngressRule{
-		{
-			Host: getPrometheusHost(cr),
-			IngressRuleValue: networkingv1.IngressRuleValue{
-				HTTP: &networkingv1.HTTPIngressRuleValue{
-					Paths: []networkingv1.HTTPIngressPath{
-						{
-							Path: getPathOrDefault(cr.Spec.Prometheus.Ingress.Path),
-							Backend: networkingv1.IngressBackend{
-								Service: &networkingv1.IngressServiceBackend{
-									Name: "prometheus-operated",
-									Port: networkingv1.ServiceBackendPort{
-										Name: "web",
-									},
-								},
-							},
-							PathType: &pathType,
-						},
-					},
-				},
-			},
-		},
-	}
-
-	// Add TLS options
-	ingress.Spec.TLS = []networkingv1.IngressTLS{
-		{
-			Hosts:      []string{cr.Name},
-			SecretName: common.ArgoCDSecretName,
-		},
-	}
-
-	// Allow override of TLS options if specified
-	if len(cr.Spec.Prometheus.Ingress.TLS) > 0 {
-		ingress.Spec.TLS = cr.Spec.Prometheus.Ingress.TLS
-	}
-
-	if err := controllerutil.SetControllerReference(cr, ingress, r.Scheme); err != nil {
-		return err
-	}
-	return r.Client.Create(context.TODO(), ingress)
-}
