--- controllers/argocd/route.go
@@ -1,125 +1,0 @@
-// Copyright 2019 ArgoCD Operator Developers
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// 	http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package argocd
-
-import (
-	"context"
-
-	routev1 "github.com/openshift/api/route/v1"
-	"k8s.io/apimachinery/pkg/util/intstr"
-	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
-
-	argoproj "github.com/argoproj-labs/argocd-operator/api/v1beta1"
-	"github.com/argoproj-labs/argocd-operator/pkg/argoutil"
-)
-
-// reconcileRoutes will ensure that all ArgoCD Routes are present.
-func (r *ReconcileArgoCD) reconcileRoutes(cr *argoproj.ArgoCD) error {
-	if err := r.reconcileGrafanaRoute(cr); err != nil {
-		return err
-	}
-
-	if err := r.reconcilePrometheusRoute(cr); err != nil {
-		return err
-	}
-
-	if err := r.reconcileServerRoute(cr); err != nil {
-		return err
-	}
-
-	if err := r.reconcileApplicationSetControllerWebhookRoute(cr); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// reconcileGrafanaRoute will ensure that the ArgoCD Grafana Route is present.
-func (r *ReconcileArgoCD) reconcileGrafanaRoute(cr *argoproj.ArgoCD) error {
-	route := newRouteWithSuffix("grafana", cr)
-	if argoutil.IsObjectFound(r.Client, cr.Namespace, route.Name, route) {
-		if !cr.Spec.Grafana.Enabled || !cr.Spec.Grafana.Route.Enabled {
-			// Route exists but enabled flag has been set to false, delete the Route
-			return r.Client.Delete(context.TODO(), route)
-		}
-		log.Info(grafanaDeprecatedWarning)
-		return nil // Route found, do nothing
-	}
-
-	if !cr.Spec.Grafana.Enabled || !cr.Spec.Grafana.Route.Enabled {
-		return nil // Grafana itself or Route not enabled, do nothing.
-	}
-
-	log.Info(grafanaDeprecatedWarning)
-
-	return nil
-}
-
-// reconcilePrometheusRoute will ensure that the ArgoCD Prometheus Route is present.
-func (r *ReconcileArgoCD) reconcilePrometheusRoute(cr *argoproj.ArgoCD) error {
-	route := newRouteWithSuffix("prometheus", cr)
-	if argoutil.IsObjectFound(r.Client, cr.Namespace, route.Name, route) {
-		if !cr.Spec.Prometheus.Enabled || !cr.Spec.Prometheus.Route.Enabled {
-			// Route exists but enabled flag has been set to false, delete the Route
-			return r.Client.Delete(context.TODO(), route)
-		}
-		return nil // Route found, do nothing
-	}
-
-	if !cr.Spec.Prometheus.Enabled || !cr.Spec.Prometheus.Route.Enabled {
-		return nil // Prometheus itself or Route not enabled, do nothing.
-	}
-
-	// Allow override of the Annotations for the Route.
-	if len(cr.Spec.Prometheus.Route.Annotations) > 0 {
-		route.Annotations = cr.Spec.Prometheus.Route.Annotations
-	}
-
-	// Allow override of the Labels for the Route.
-	if len(cr.Spec.Prometheus.Route.Labels) > 0 {
-		labels := route.Labels
-		for key, val := range cr.Spec.Prometheus.Route.Labels {
-			labels[key] = val
-		}
-		route.Labels = labels
-	}
-
-	// Allow override of the Host for the Route.
-	if len(cr.Spec.Prometheus.Host) > 0 {
-		route.Spec.Host = cr.Spec.Prometheus.Host // TODO: What additional role needed for this?
-	}
-
-	route.Spec.Port = &routev1.RoutePort{
-		TargetPort: intstr.FromString("web"),
-	}
-
-	// Allow override of TLS options for the Route
-	if cr.Spec.Prometheus.Route.TLS != nil {
-		route.Spec.TLS = cr.Spec.Prometheus.Route.TLS
-	}
-
-	route.Spec.To.Kind = "Service"
-	route.Spec.To.Name = "prometheus-operated"
-
-	// Allow override of the WildcardPolicy for the Route
-	if cr.Spec.Prometheus.Route.WildcardPolicy != nil && len(*cr.Spec.Prometheus.Route.WildcardPolicy) > 0 {
-		route.Spec.WildcardPolicy = *cr.Spec.Prometheus.Route.WildcardPolicy
-	}
-
-	if err := controllerutil.SetControllerReference(cr, route, r.Scheme); err != nil {
-		return err
-	}
-	return r.Client.Create(context.TODO(), route)
-}
